refactor: moved JsonHelper and FallbackWriter out of the journal package
