+++ bugsnag-android-core/src/main/java/com/bugsnag/android/Client.java
@@ -179,7 +179,8 @@ public class Client implements MetadataAware, CallbackAware, UserAware {
         registerLifecycleCallbacks();
 
         EventStorageModule eventStorageModule = new EventStorageModule(contextModule, configModule,
-                dataCollectionModule, bgTaskService, trackerModule, systemServiceModule, notifier);
+                dataCollectionModule, bgTaskService, trackerModule, systemServiceModule, notifier,
+                callbackState);
         eventStorageModule.resolveDependencies(bgTaskService, TaskType.IO);
         eventStore = eventStorageModule.getEventStore();
