+++ bugsnag-android-core/src/test/java/com/bugsnag/android/LaunchCrashDeliveryTest.kt
@@ -159,7 +159,8 @@ class LaunchCrashDeliveryTest {
             NoopLogger,
             Notifier(),
             backgroundTaskService,
-            FileStore.Delegate { _, _, _ -> }
+            FileStore.Delegate { _, _, _ -> },
+            CallbackState()
         )
     }
 }
